filter perf cases
