+++ driver/driver.hpp
@@ -151,8 +151,7 @@ inline void PadBufferSize(size_t& sz, int datatype_sz)
            "pool[fp16], lrn[fp16], "
            "activ[fp16], softmax[fp16], bnorm[fp16], rnn[fp16], gemm[fp16], ctc, dropout[fp16], "
            "tensorop[fp16], reduce[fp16|fp64], layernorm[bfp16|fp16], sum[bfp16|fp16], "
-           "argmax[bfp16|fp16], groupnorm[bfp16|fp16], cat[bfp16|fp16] diag[bfp16|fp16] "
-           "diagflat[bfp16|fp16] diagembed[bfp16|fp16]\n");
+           "argmax[bfp16|fp16], groupnorm[bfp16|fp16], cat[bfp16|fp16], diag[bfp16|fp16]\n");
     exit(0); // NOLINT (concurrency-mt-unsafe)
 }
 
@@ -178,9 +177,7 @@ inline std::string ParseBaseArg(int argc, char* argv[])
        arg != "sumbfp16" && arg != "argmax" && arg != "argmaxfp16" && arg != "argmaxbfp16" &&
        arg != "groupnorm" && arg != "groupnormfp16" && arg != "groupnormbfp16" && arg != "cat" &&
        arg != "catfp16" && arg != "catbfp16" && arg != "diag" && arg != "diagfp16" &&
-       arg != "diagbfp16" && arg != "diagflat" && arg != "diagflatfp16" && arg != "diagflatbfp16" &&
-       arg != "diagembed" && arg != "diagembedfp16" && arg != "diagembedbfp16" &&
-       arg != "--version")
+       arg != "diagbfp16" && arg != "--version")
     {
         printf("FAILED: Invalid Base Input Argument\n");
         Usage();
