--- src/include/miopen/diagonal/diag/problem_description.hpp
@@ -1,98 +1,0 @@
-/*******************************************************************************
- *
- * MIT License
- *
- * Copyright (c) 2024 Advanced Micro Devices, Inc.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- *
- *******************************************************************************/
-
-#pragma once
-
-#include "miopen/tensor_view_utils.hpp"
-#include <cstdint>
-#include <miopen/problem_description_base.hpp>
-#include <miopen/tensor.hpp>
-#include <miopen/activ.hpp>
-
-#include <string>
-
-namespace miopen {
-
-struct NetworkConfig;
-
-namespace diagonal {
-
-namespace diag {
-
-struct FwdProblemDescription : ProblemDescriptionBase
-{
-    // Forward constructor
-    FwdProblemDescription(const TensorDescriptor& inputDesc_,
-                          const TensorDescriptor& outputDesc_,
-                          int64_t diagonal_)
-        : inputDesc(inputDesc_), outputDesc(outputDesc_), diagonal(diagonal_)
-    {
-        if(inputDesc.GetSize() != 1 && inputDesc.GetSize() != 2)
-        {
-
-            MIOPEN_THROW(miopenStatusBadParm,
-                         "Diag::FwdProblemDescription: Number of tensor dimension must be 1 or 2.");
-        }
-    }
-
-    const TensorDescriptor& GetInputDesc() const { return inputDesc; }
-    const TensorDescriptor& GetOutputDesc() const { return outputDesc; }
-    int64_t GetDiagonal() const { return diagonal; }
-
-    bool IsSameType() const
-    {
-        if(inputDesc.GetType() != outputDesc.GetType())
-        {
-            return false;
-        }
-
-        return true;
-    }
-
-    bool IsAllPacked() const
-    {
-        if(!(inputDesc.IsPacked() && outputDesc.IsPacked()))
-        {
-            return false;
-        }
-
-        return true;
-    }
-
-    NetworkConfig MakeNetworkConfig() const override;
-
-private:
-    TensorDescriptor inputDesc;
-    TensorDescriptor outputDesc;
-
-    int64_t diagonal;
-};
-
-} // namespace diag
-
-} // namespace diagonal
-
-} // namespace miopen
